refactor(work-item-lifecycle): share forge repository adapters
Hoist the GitHub/GitLab/Azure DevOps adapters next to forgeDisplayName so
Create PR and Mark PR Ready for Review use one copy.
